Messaging/final_project/requirements.txt
requests
pika
sseclient
psycopg2-binary